Improve server side git url handling and server test improvements (#2237)
