+++ build.gradle
@@ -143,8 +143,8 @@ jacocoTestCoverageVerification {
             limit {
                 counter = "LINE"
                 value = "COVEREDRATIO"
-                // TODO: in the future the following value should become at least 0.80
-                minimum = 0.79
+                // TODO: in the future the following value should become at least 0.90
+                minimum = 0.80
             }
         }
     }
