+++ src/main/resources/templates/kotlin/test/projectTemplate/pom.xml
@@ -13,13 +13,13 @@
     <name>${exerciseName} Tests</name>
 	<properties>
 		<project.build.sourceEncoding>UTF-8</project.build.sourceEncoding>
-        <kotlin.version>1.3.50</kotlin.version>
+        <kotlin.version>1.4.10</kotlin.version>
 	</properties>
     <dependencies>
         <dependency>
             <groupId>junit</groupId>
             <artifactId>junit</artifactId>
-            <version>4.12</version>
+            <version>4.13.1</version>
         </dependency>
         <dependency>
             <groupId>org.json</groupId>
@@ -109,8 +109,8 @@
                     </execution>
                 </executions>
                 <configuration>
-                    <source>12</source>
-                    <target>12</target>
+                    <source>15</source>
+                    <target>15</target>
                 </configuration>
             </plugin>
         </plugins>
