+++ src/main/webapp/app/exercises/programming/manage/programming-exercise-detail.component.html
@@ -124,7 +124,7 @@
                     <span>{{ programmingExercise.staticCodeAnalysisEnabled }}</span>
                 </dd>
                 <ng-container *ngIf="programmingExercise.staticCodeAnalysisEnabled">
-                    <dt><span jhiTranslate="artemisApp.programmingExercise.maxStaticCodeAnalysisPenalty">Max Static Code Analysis Penalty</span></dt>
+                    <dt><span jhiTranslate="artemisApp.programmingExercise.maxStaticCodeAnalysisPenalty.title">Max Static Code Analysis Penalty</span></dt>
                     <dd>
                         <span>{{ programmingExercise.maxStaticCodeAnalysisPenalty }}</span>
                     </dd>
