+++ src/test/java/de/tum/in/www1/artemis/AbstractSpringIntegrationBambooBitbucketJiraTest.java
@@ -7,40 +7,42 @@ import static org.mockito.ArgumentMatchers.any;
 import static org.mockito.Mockito.doReturn;
 
 import java.io.IOException;
+import java.net.MalformedURLException;
 import java.net.URISyntaxException;
-import java.util.*;
+import java.util.ArrayList;
+import java.util.LinkedList;
+import java.util.List;
+import java.util.Set;
 
 import org.junit.jupiter.api.AfterEach;
 import org.mockito.Mockito;
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.beans.factory.annotation.Value;
 import org.springframework.boot.test.autoconfigure.jdbc.AutoConfigureTestDatabase;
 import org.springframework.boot.test.autoconfigure.web.servlet.AutoConfigureMockMvc;
 import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.boot.test.mock.mockito.SpyBean;
-import org.springframework.messaging.simp.SimpMessageSendingOperations;
 import org.springframework.test.context.ActiveProfiles;
 
 import com.atlassian.bamboo.specs.util.BambooServer;
+import com.fasterxml.jackson.core.JsonProcessingException;
 
 import de.tum.in.www1.artemis.connector.bamboo.BambooRequestMockProvider;
 import de.tum.in.www1.artemis.connector.bitbucket.BitbucketRequestMockProvider;
 import de.tum.in.www1.artemis.domain.ProgrammingExercise;
+import de.tum.in.www1.artemis.domain.Team;
 import de.tum.in.www1.artemis.domain.User;
 import de.tum.in.www1.artemis.domain.enumeration.RepositoryType;
-import de.tum.in.www1.artemis.service.*;
+import de.tum.in.www1.artemis.domain.participation.ProgrammingExerciseStudentParticipation;
 import de.tum.in.www1.artemis.service.connectors.BitbucketBambooUpdateService;
-import de.tum.in.www1.artemis.service.connectors.GitService;
-import de.tum.in.www1.artemis.service.connectors.LtiService;
 import de.tum.in.www1.artemis.service.connectors.bamboo.BambooService;
+import de.tum.in.www1.artemis.service.connectors.bamboo.dto.BambooBuildResultDTO;
 import de.tum.in.www1.artemis.service.connectors.bamboo.dto.BambooRepositoryDTO;
 import de.tum.in.www1.artemis.service.connectors.bamboo.dto.BambooTriggerDTO;
 import de.tum.in.www1.artemis.service.connectors.bitbucket.BitbucketService;
 import de.tum.in.www1.artemis.service.connectors.bitbucket.dto.BitbucketRepositoryDTO;
 import de.tum.in.www1.artemis.service.ldap.LdapUserService;
-import de.tum.in.www1.artemis.service.messaging.InstanceMessageSendService;
-import de.tum.in.www1.artemis.service.scheduled.ProgrammingExerciseScheduleService;
-import de.tum.in.www1.artemis.util.*;
+import de.tum.in.www1.artemis.util.AbstractArtemisIntegrationTest;
+import de.tum.in.www1.artemis.util.Verifiable;
 
 @SpringBootTest(properties = { "artemis.athene.token-validity-in-seconds=10800",
         "artemis.athene.base64-secret=YWVuaXF1YWRpNWNlaXJpNmFlbTZkb283dXphaVF1b29oM3J1MWNoYWlyNHRoZWUzb2huZ2FpM211bGVlM0VpcAo=" })
@@ -48,17 +50,10 @@ import de.tum.in.www1.artemis.util.*;
 @AutoConfigureTestDatabase
 // NOTE: we use a common set of active profiles to reduce the number of application launches during testing. This significantly saves time and memory!
 @ActiveProfiles({ "artemis", "bamboo", "bitbucket", "jira", "automaticText", "ldap", "scheduling", "athene" })
-public abstract class AbstractSpringIntegrationBambooBitbucketJiraTest {
-
-    @Value("${server.url}")
-    protected String ARTEMIS_SERVER_URL;
+public abstract class AbstractSpringIntegrationBambooBitbucketJiraTest extends AbstractArtemisIntegrationTest {
 
     @SpyBean
     protected LdapUserService ldapUserService;
-
-    // NOTE: we prefer SpyBean over MockBean, because it is more lightweight, we can mock method, but we can also invoke actual methods during testing
-    @SpyBean
-    protected LtiService ltiService;
 
     // please only use this to verify method calls using Mockito. Do not mock methods, instead mock the communication with Bamboo using the corresponding RestTemplate.
     @SpyBean
@@ -75,56 +70,20 @@ public abstract class AbstractSpringIntegrationBambooBitbucketJiraTest {
     @SpyBean
     protected BambooServer bambooServer;
 
-    @SpyBean
-    protected GitService gitService;
-
-    @SpyBean
-    protected GroupNotificationService groupNotificationService;
-
-    @SpyBean
-    protected WebsocketMessagingService websocketMessagingService;
-
-    @SpyBean
-    protected PlantUmlService plantUmlService;
-
-    @SpyBean
-    protected SimpMessageSendingOperations messagingTemplate;
-
-    @SpyBean
-    protected ProgrammingSubmissionService programmingSubmissionService;
-
-    @SpyBean
-    protected ExamAccessService examAccessService;
-
-    @SpyBean
-    protected InstanceMessageSendService instanceMessageSendService;
-
-    @SpyBean
-    protected ProgrammingExerciseScheduleService programmingExerciseScheduleService;
-
-    @SpyBean
-    protected ProgrammingExerciseParticipationService programmingExerciseParticipationServiceSpy;
-
     @Autowired
     protected BambooRequestMockProvider bambooRequestMockProvider;
 
     @Autowired
     protected BitbucketRequestMockProvider bitbucketRequestMockProvider;
 
-    @Autowired
-    protected DatabaseUtilService database;
-
-    @Autowired
-    protected RequestUtilService request;
-
     @AfterEach
     public void resetSpyBeans() {
-        Mockito.reset(ltiService, continuousIntegrationService, versionControlService, bambooServer, gitService, groupNotificationService, websocketMessagingService,
-                plantUmlService, messagingTemplate, programmingSubmissionService, examAccessService, instanceMessageSendService, programmingExerciseScheduleService,
-                programmingExerciseParticipationServiceSpy);
+        Mockito.reset(ldapUserService, continuousIntegrationUpdateService, continuousIntegrationService, versionControlService, bambooServer);
+        super.resetSpyBeans();
     }
 
-    protected List<Verifiable> mockConnectorRequestsForStartParticipation(ProgrammingExercise exercise, String username, Set<User> users) throws IOException, URISyntaxException {
+    @Override
+    public List<Verifiable> mockConnectorRequestsForStartParticipation(ProgrammingExercise exercise, String username, Set<User> users) throws IOException, URISyntaxException {
         final var verifications = new LinkedList<Verifiable>();
         bitbucketRequestMockProvider.mockCopyRepositoryForParticipation(exercise, username);
         bitbucketRequestMockProvider.mockConfigureRepository(exercise, username, users);
@@ -135,14 +94,16 @@ public abstract class AbstractSpringIntegrationBambooBitbucketJiraTest {
         return verifications;
     }
 
-    protected void mockUpdatePlanRepositoryForParticipation(ProgrammingExercise exercise, String username) throws IOException, URISyntaxException {
+    @Override
+    public void mockUpdatePlanRepositoryForParticipation(ProgrammingExercise exercise, String username) throws IOException, URISyntaxException {
         final var projectKey = exercise.getProjectKey();
         final var bitbucketRepoName = projectKey.toLowerCase() + "-" + username;
 
         mockUpdatePlanRepository(exercise, username, ASSIGNMENT_REPO_NAME, bitbucketRepoName, List.of());
     }
 
-    private void mockUpdatePlanRepository(ProgrammingExercise exercise, String planName, String bambooRepoName, String bitbucketRepoName, List<String> triggeredBy)
+    @Override
+    public void mockUpdatePlanRepository(ProgrammingExercise exercise, String planName, String bambooRepoName, String bitbucketRepoName, List<String> triggeredBy)
             throws IOException, URISyntaxException {
         final var projectKey = exercise.getProjectKey();
         final var buildPlanKey = (projectKey + "-" + planName).toUpperCase();
@@ -181,7 +142,8 @@ public abstract class AbstractSpringIntegrationBambooBitbucketJiraTest {
         }
     }
 
-    protected void mockConnectorRequestsForSetup(ProgrammingExercise exercise) throws IOException, URISyntaxException {
+    @Override
+    public void mockConnectorRequestsForSetup(ProgrammingExercise exercise) throws Exception {
         final var projectKey = exercise.getProjectKey();
         String exerciseRepoName = projectKey.toLowerCase() + "-" + RepositoryType.TEMPLATE.getName();
         String testRepoName = projectKey.toLowerCase() + "-" + RepositoryType.TESTS.getName();
@@ -200,7 +162,8 @@ public abstract class AbstractSpringIntegrationBambooBitbucketJiraTest {
         doReturn(null).when(bambooServer).publish(any());
     }
 
-    protected List<Verifiable> mockConnectorRequestsForImport(ProgrammingExercise sourceExercise, ProgrammingExercise exerciseToBeImported) throws IOException, URISyntaxException {
+    @Override
+    public List<Verifiable> mockConnectorRequestsForImport(ProgrammingExercise sourceExercise, ProgrammingExercise exerciseToBeImported) throws IOException, URISyntaxException {
         final var verifications = new ArrayList<Verifiable>();
         final var projectKey = exerciseToBeImported.getProjectKey();
         final var sourceProjectKey = sourceExercise.getProjectKey();
@@ -239,51 +202,35 @@ public abstract class AbstractSpringIntegrationBambooBitbucketJiraTest {
         return verifications;
     }
 
-    /**
-     * can be invoked for teams and students
-     */
-    protected void setupRepositoryMocksParticipant(ProgrammingExercise exercise, String participantName, LocalRepository studentRepo) throws Exception {
-        final var projectKey = exercise.getProjectKey();
-        String participantRepoName = projectKey.toLowerCase() + "-" + participantName;
-        var participantRepoTestUrl = new GitUtilService.MockFileRepositoryUrl(studentRepo.originRepoFile);
-        doReturn(participantRepoTestUrl).when(versionControlService).getCloneRepositoryUrl(projectKey, participantRepoName);
-        doReturn(gitService.getRepositoryByLocalPath(studentRepo.localRepoFile.toPath())).when(gitService).getOrCheckoutRepository(participantRepoTestUrl.getURL(), true);
-        doReturn(participantRepoName).when(continuousIntegrationService).getRepositorySlugFromUrl(participantRepoTestUrl.getURL());
-        doReturn(projectKey).when(continuousIntegrationService).getProjectKeyFromUrl(participantRepoTestUrl.getURL());
-        doReturn(participantRepoName).when(versionControlService).getRepositorySlugFromUrl(participantRepoTestUrl.getURL());
+    @Override
+    public void mockRemoveRepositoryAccess(ProgrammingExercise exercise, Team team, User firstStudent) throws URISyntaxException {
+        final var repositorySlug = (exercise.getProjectKey() + "-" + team.getParticipantIdentifier()).toLowerCase();
+        bitbucketRequestMockProvider.mockRemoveMemberFromRepository(repositorySlug, exercise.getProjectKey(), firstStudent);
     }
 
-    protected void setupRepositoryMocks(ProgrammingExercise exercise, LocalRepository exerciseRepo, LocalRepository solutionRepo, LocalRepository testRepo) throws Exception {
-        final var projectKey = exercise.getProjectKey();
+    @Override
+    public void mockRepositoryWritePermissions(Team team, User newStudent, ProgrammingExercise exercise) throws URISyntaxException {
+        final var repositorySlug = (exercise.getProjectKey() + "-" + team.getParticipantIdentifier()).toLowerCase();
+        bitbucketRequestMockProvider.mockGiveWritePermission(exercise, repositorySlug, newStudent.getLogin());
+    }
 
-        String exerciseRepoName = projectKey.toLowerCase() + "-" + RepositoryType.TEMPLATE.getName();
-        String testRepoName = projectKey.toLowerCase() + "-" + RepositoryType.TESTS.getName();
-        String solutionRepoName = projectKey.toLowerCase() + "-" + RepositoryType.SOLUTION.getName();
+    @Override
+    public void mockRetrieveArtifacts(ProgrammingExerciseStudentParticipation participation) throws MalformedURLException, URISyntaxException, JsonProcessingException {
+        // prepare the build result
+        bambooRequestMockProvider.mockQueryLatestBuildResultFromBambooServer(participation.getBuildPlanId());
+        // prepare the artifact to be null
+        bambooRequestMockProvider.mockRetrieveEmptyArtifactPage();
+    }
 
-        var exerciseRepoTestUrl = new GitUtilService.MockFileRepositoryUrl(exerciseRepo.originRepoFile);
-        var testRepoTestUrl = new GitUtilService.MockFileRepositoryUrl(testRepo.originRepoFile);
-        var solutionRepoTestUrl = new GitUtilService.MockFileRepositoryUrl(solutionRepo.originRepoFile);
+    @Override
+    public void mockGetBuildLogs(ProgrammingExerciseStudentParticipation participation, List<BambooBuildResultDTO.BambooBuildLogEntryDTO> logs)
+            throws URISyntaxException, JsonProcessingException {
+        bambooRequestMockProvider.mockGetBuildLogs(participation.getBuildPlanId(), logs);
+    }
 
-        doReturn(exerciseRepoTestUrl).when(versionControlService).getCloneRepositoryUrl(projectKey, exerciseRepoName);
-        doReturn(testRepoTestUrl).when(versionControlService).getCloneRepositoryUrl(projectKey, testRepoName);
-        doReturn(solutionRepoTestUrl).when(versionControlService).getCloneRepositoryUrl(projectKey, solutionRepoName);
-
-        doReturn(gitService.getRepositoryByLocalPath(exerciseRepo.localRepoFile.toPath())).when(gitService).getOrCheckoutRepository(exerciseRepoTestUrl.getURL(), true);
-        doReturn(gitService.getRepositoryByLocalPath(testRepo.localRepoFile.toPath())).when(gitService).getOrCheckoutRepository(testRepoTestUrl.getURL(), true);
-        doReturn(gitService.getRepositoryByLocalPath(solutionRepo.localRepoFile.toPath())).when(gitService).getOrCheckoutRepository(solutionRepoTestUrl.getURL(), true);
-
-        doReturn(exerciseRepoName).when(continuousIntegrationService).getRepositorySlugFromUrl(exerciseRepoTestUrl.getURL());
-        doReturn(testRepoName).when(continuousIntegrationService).getRepositorySlugFromUrl(testRepoTestUrl.getURL());
-        doReturn(solutionRepoName).when(continuousIntegrationService).getRepositorySlugFromUrl(solutionRepoTestUrl.getURL());
-
-        doReturn(projectKey).when(continuousIntegrationService).getProjectKeyFromUrl(exerciseRepoTestUrl.getURL());
-        doReturn(projectKey).when(continuousIntegrationService).getProjectKeyFromUrl(testRepoTestUrl.getURL());
-        doReturn(projectKey).when(continuousIntegrationService).getProjectKeyFromUrl(solutionRepoTestUrl.getURL());
-
-        doReturn(exerciseRepoName).when(versionControlService).getRepositorySlugFromUrl(exerciseRepoTestUrl.getURL());
-        doReturn(testRepoName).when(versionControlService).getRepositorySlugFromUrl(testRepoTestUrl.getURL());
-        doReturn(solutionRepoName).when(versionControlService).getRepositorySlugFromUrl(solutionRepoTestUrl.getURL());
-
-        doReturn(projectKey).when(versionControlService).getProjectKeyFromUrl(any());
+    @Override
+    public void resetMockProvider() {
+        bitbucketRequestMockProvider.reset();
+        bambooRequestMockProvider.reset();
     }
 }
