+++ src/test/java/de/tum/in/www1/artemis/AbstractSpringIntegrationJenkinsGitlabTest.java
@@ -1,5 +1,11 @@
 package de.tum.in.www1.artemis;
 
+import java.io.IOException;
+import java.net.MalformedURLException;
+import java.net.URISyntaxException;
+import java.util.List;
+import java.util.Set;
+
 import org.junit.jupiter.api.AfterEach;
 import org.mockito.Mockito;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -7,22 +13,24 @@ import org.springframework.boot.test.autoconfigure.jdbc.AutoConfigureTestDatabas
 import org.springframework.boot.test.autoconfigure.web.servlet.AutoConfigureMockMvc;
 import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.boot.test.mock.mockito.SpyBean;
-import org.springframework.messaging.simp.SimpMessageSendingOperations;
 import org.springframework.test.context.ActiveProfiles;
 import org.springframework.test.context.TestPropertySource;
 
+import com.fasterxml.jackson.core.JsonProcessingException;
 import com.offbytwo.jenkins.JenkinsServer;
 
-import de.tum.in.www1.artemis.service.GroupNotificationService;
-import de.tum.in.www1.artemis.service.PlantUmlService;
-import de.tum.in.www1.artemis.service.ProgrammingSubmissionService;
-import de.tum.in.www1.artemis.service.WebsocketMessagingService;
-import de.tum.in.www1.artemis.service.connectors.GitService;
-import de.tum.in.www1.artemis.service.connectors.LtiService;
+import de.tum.in.www1.artemis.connector.gitlab.GitlabRequestMockProvider;
+import de.tum.in.www1.artemis.connector.jenkins.JenkinsRequestMockProvider;
+import de.tum.in.www1.artemis.domain.ProgrammingExercise;
+import de.tum.in.www1.artemis.domain.Team;
+import de.tum.in.www1.artemis.domain.User;
+import de.tum.in.www1.artemis.domain.enumeration.RepositoryType;
+import de.tum.in.www1.artemis.domain.participation.ProgrammingExerciseStudentParticipation;
+import de.tum.in.www1.artemis.service.connectors.bamboo.dto.BambooBuildResultDTO;
 import de.tum.in.www1.artemis.service.connectors.gitlab.GitLabService;
 import de.tum.in.www1.artemis.service.connectors.jenkins.JenkinsService;
-import de.tum.in.www1.artemis.util.DatabaseUtilService;
-import de.tum.in.www1.artemis.util.RequestUtilService;
+import de.tum.in.www1.artemis.util.AbstractArtemisIntegrationTest;
+import de.tum.in.www1.artemis.util.Verifiable;
 
 @SpringBootTest
 @AutoConfigureMockMvc
@@ -30,11 +38,7 @@ import de.tum.in.www1.artemis.util.RequestUtilService;
 // NOTE: we use a common set of active profiles to reduce the number of application launches during testing. This significantly saves time and memory!
 @ActiveProfiles({ "artemis", "gitlab", "jenkins", "automaticText" })
 @TestPropertySource(properties = "artemis.user-management.use-external=false")
-public abstract class AbstractSpringIntegrationJenkinsGitlabTest {
-
-    // NOTE: we prefer SpyBean over MockBean, because it is more lightweight, we can mock method, but we can also invoke actual methods during testing
-    @SpyBean
-    protected LtiService ltiService;
+public abstract class AbstractSpringIntegrationJenkinsGitlabTest extends AbstractArtemisIntegrationTest {
 
     // please only use this to verify method calls using Mockito. Do not mock methods, instead mock the communication with Jenkins using the corresponding RestTemplate.
     @SpyBean
@@ -48,33 +52,82 @@ public abstract class AbstractSpringIntegrationJenkinsGitlabTest {
     @SpyBean
     protected JenkinsServer jenkinsServer;
 
-    @SpyBean
-    protected GitService gitService;
-
-    @SpyBean
-    protected GroupNotificationService groupNotificationService;
-
-    @SpyBean
-    protected WebsocketMessagingService websocketMessagingService;
-
-    @SpyBean
-    protected PlantUmlService plantUmlService;
-
-    @SpyBean
-    protected SimpMessageSendingOperations messagingTemplate;
-
-    @SpyBean
-    protected ProgrammingSubmissionService programmingSubmissionService;
+    @Autowired
+    protected JenkinsRequestMockProvider jenkinsRequestMockProvider;
 
     @Autowired
-    protected DatabaseUtilService database;
-
-    @Autowired
-    protected RequestUtilService request;
+    protected GitlabRequestMockProvider gitlabRequestMockProvider;
 
     @AfterEach
     public void resetSpyBeans() {
-        Mockito.reset(ltiService, continuousIntegrationService, versionControlService, jenkinsServer, gitService, groupNotificationService, websocketMessagingService,
-                plantUmlService, messagingTemplate, programmingSubmissionService);
+        Mockito.reset(continuousIntegrationService, versionControlService, jenkinsServer);
+        super.resetSpyBeans();
+    }
+
+    @Override
+    public void mockConnectorRequestsForSetup(ProgrammingExercise exercise) throws Exception {
+        final var projectKey = exercise.getProjectKey();
+        String exerciseRepoName = projectKey.toLowerCase() + "-" + RepositoryType.TEMPLATE.getName();
+        String testRepoName = projectKey.toLowerCase() + "-" + RepositoryType.TESTS.getName();
+        String solutionRepoName = projectKey.toLowerCase() + "-" + RepositoryType.SOLUTION.getName();
+        gitlabRequestMockProvider.mockCheckIfProjectExists(exercise, false);
+        gitlabRequestMockProvider.mockCreateProjectForExercise(exercise);
+        gitlabRequestMockProvider.mockCreateRepository(exercise, exerciseRepoName);
+        gitlabRequestMockProvider.mockCreateRepository(exercise, testRepoName);
+        gitlabRequestMockProvider.mockCreateRepository(exercise, solutionRepoName);
+        gitlabRequestMockProvider.mockAddAuthenticatedWebHook();
+        jenkinsRequestMockProvider.mockCreateProjectForExercise(exercise);
+        jenkinsRequestMockProvider.mockCreateBuildPlan(projectKey);
+        jenkinsRequestMockProvider.mockTriggerBuild();
+    }
+
+    @Override
+    public List<Verifiable> mockConnectorRequestsForImport(ProgrammingExercise sourceExercise, ProgrammingExercise exerciseToBeImported) throws IOException, URISyntaxException {
+        // TODO: implement
+        return null;
+    }
+
+    @Override
+    public List<Verifiable> mockConnectorRequestsForStartParticipation(ProgrammingExercise exercise, String username, Set<User> users) throws IOException, URISyntaxException {
+        // TODO: implement
+        return null;
+    }
+
+    @Override
+    public void mockUpdatePlanRepositoryForParticipation(ProgrammingExercise exercise, String username) throws IOException, URISyntaxException {
+        // TODO: implement
+    }
+
+    @Override
+    public void mockUpdatePlanRepository(ProgrammingExercise exercise, String planName, String bambooRepoName, String bitbucketRepoName, List<String> triggeredBy)
+            throws IOException, URISyntaxException {
+        // TODO: implement
+    }
+
+    @Override
+    public void mockRemoveRepositoryAccess(ProgrammingExercise exercise, Team team, User firstStudent) throws URISyntaxException {
+        // TODO: implement
+    }
+
+    @Override
+    public void mockRepositoryWritePermissions(Team team, User newStudent, ProgrammingExercise exercise) throws URISyntaxException {
+        // TODO: implement
+    }
+
+    @Override
+    public void mockRetrieveArtifacts(ProgrammingExerciseStudentParticipation participation) throws MalformedURLException, URISyntaxException, JsonProcessingException {
+        // TODO: implement
+    }
+
+    @Override
+    public void mockGetBuildLogs(ProgrammingExerciseStudentParticipation participation, List<BambooBuildResultDTO.BambooBuildLogEntryDTO> logs)
+            throws URISyntaxException, JsonProcessingException {
+        // TODO: implement
+    }
+
+    @Override
+    public void resetMockProvider() {
+        gitlabRequestMockProvider.reset();
+        jenkinsRequestMockProvider.reset();
     }
 }
