+++ src/test/java/de/tum/in/www1/artemis/programmingexercise/ProgrammingExerciseResultJenkinsIntegrationTest.java
@@ -1,0 +1,45 @@
+package de.tum.in.www1.artemis.programmingexercise;
+
+import java.io.IOException;
+import java.util.List;
+
+import org.junit.jupiter.api.AfterEach;
+import org.junit.jupiter.api.BeforeEach;
+import org.junit.jupiter.api.Test;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.security.test.context.support.WithMockUser;
+
+import de.tum.in.www1.artemis.AbstractSpringIntegrationJenkinsGitlabTest;
+import de.tum.in.www1.artemis.config.Constants;
+import de.tum.in.www1.artemis.util.ModelFactory;
+import de.tum.in.www1.artemis.util.ProgrammingExerciseResultTestService;
+
+class ProgrammingExerciseResultJenkinsIntegrationTest extends AbstractSpringIntegrationJenkinsGitlabTest {
+
+    @Autowired
+    private ProgrammingExerciseResultTestService programmingExerciseResultTestService;
+
+    @BeforeEach
+    void setup() throws Exception {
+        programmingExerciseResultTestService.setup();
+    }
+
+    @AfterEach
+    void tearDown() throws IOException {
+        programmingExerciseResultTestService.tearDown();
+    }
+
+    @Test
+    @WithMockUser(value = "student1", roles = "USER")
+    public void shouldUpdateTestCasesAndResultScoreFromSolutionParticipationResult() throws Exception {
+        var notification = ModelFactory.generateTestResultDTO(Constants.ASSIGNMENT_REPO_NAME, List.of("test1", "test2", "test4"), List.of());
+        programmingExerciseResultTestService.shouldUpdateTestCasesAndResultScoreFromSolutionParticipationResult(notification);
+    }
+
+    @Test
+    @WithMockUser(value = "student1", roles = "USER")
+    public void shouldStoreFeedbackForResultWithStaticCodeAnalysisReport() throws Exception {
+        var notification = ModelFactory.generateTestResultDTO(Constants.ASSIGNMENT_REPO_NAME, List.of("test1"), List.of());
+        programmingExerciseResultTestService.shouldStoreFeedbackForResultWithStaticCodeAnalysisReport(notification);
+    }
+}
